1.3.0 Dev Preview 1
Reworked the calls for GML Functions, added a universal function to make it a lot easier.
Added some features to the "Game" tab, which was renamed from "World", as the features that were there wasn't what I had in mind for the tab. There is no release for this version, as it is a development preview, so you'll have to build it if you want to use the new features.